+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # hawa-docs
 
+## 0.0.41
+
+### Patch Changes
+
+- Updated dependencies [d60e897]
+  - @sikka/hawa@0.30.5
+
 ## 0.0.40
 
 ### Patch Changes